+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [4.6.1] - 2023-03-30
+
+### Fixed
+
+- Update kubernetes registry from k8s.gcr.io Image to registry.k8s.io, based on this [announcement](https://kubernetes.io/blog/2023/02/06/k8s-gcr-io-freeze-announcement/)
+- Typo on Tempo Helm chart version variable name.
+- Typo on Tempo priority class name set value.
 
 ## [4.6.0] - 2023-02-23